--- a/src/main.cpp
+++ b/src/main.cpp
@@ -870,8 +870,12 @@ int main(int argc, char** argv)
 		for (auto &fn : files_list) {
 			++numFilesProcessed;
 			double time_file_start = getsec();
-
-			std::cout << "Processing file [" << numFilesProcessed << "/" << files_count << "] \"" << fn.filename() << "\":" << (verbose ? "\n" : " ");
+			printf("Processing file [%d/%d] \"%s\":%s",
+				numFilesProcessed,
+				files_count,
+				fn.filename().string().c_str(),
+				(verbose ? "\n" : " ")
+			);
 
 			try {
 #if defined(WIN32) && defined(UNICODE)
@@ -897,12 +901,12 @@ int main(int argc, char** argv)
 			int el_h = (int) elapsed / (60 * 60);
 			int el_m = (int) (elapsed - el_h * 60 * 60) / 60;
 			int el_s = (int) (elapsed - el_h * 60 * 60 - el_m * 60);
-			std::cout << "Done, took: ";
+			printf("Done, took: ");
 			if (el_h)
-				std::cout << el_h << "h";
+				printf("%dh", el_h);
 			if (el_m)
-				std::cout << el_m << "m";
-			std::cout << el_s << "s total, file: " << time_file << "s avg: " << timeAvg << "s" << std::endl;
+				printf("%dm", el_h);
+			printf("%ds total, file: %.3fs avg: %.3fs\n", el_s, time_file, timeAvg);
 		}
 
 
@@ -924,7 +928,7 @@ int main(int argc, char** argv)
 		}
 		double time_end = getsec();
 		double time_file = time_end - time_file_start;
-		std::cout << "Done, took: " << time_file << "s total, file: " << time_file << "s avg: " << time_file << "s" << std::endl;
+		printf("Done, took: %.3fs total, file: %.3fs avg: %.3fs\n", time_file, time_file, time_file);
 	}
 	
 
@@ -935,7 +939,7 @@ int main(int argc, char** argv)
 		double gflops_proc = (converter->flops.flop / (1000.0*1000.0*1000.0)) / converter->flops.filter_sec;
 		double gflops_all = (converter->flops.flop / (1000.0*1000.0*1000.0)) / (time_end - time_start);
 
-		printf("Finished processing %d files%s%fsecs total, filter: %fs. %d files skipped, %d files errored. [GFLOPS: %05f, GFLOPS-Filter: %05f]\n",
+		printf("Finished processing %d files%s%.3fsecs total, filter: %.3fsecs; %d files skipped, %d files errored. [GFLOPS: %7.2f, GFLOPS-Filter: %7.2f]\n",
 			numFilesProcessed,
 			(verbose ? "\nTook: " : ", took: "),
 			(time_end - time_start),