Block process log now shows the total rows and columns.

The log messages counting the files now says "Processing" at the start and "Done"  when finished, the instead of previouss "Ok." at the start.

Also single file operations now show the same amount of logging as do folder operations

E.g:
## Before:
Silent:
``` 
K:\w2x\waifu2x-converter-cpp\out\Release>waifu2x-converter-cpp.exe -i 7KMAtWE.jpg -s
CUDA: GeForce RTX 2060
process successfully done! (all:2.93005[sec], 1 [files processed], 0 [files errored], 1994.88[GFLOPS], filter:1.7955[sec], 3255.41[GFLOPS])
```
Verbose:
```
K:\w2x\waifu2x-converter-cpp\out\Release>waifu2x-converter-cpp.exe -i 7KMAtWE.jpg
CUDA: GeForce RTX 2060
start process block (0,0) ...
.............
.............
.............
start process block (4,6) ...
process successfully done! (all:4.2414[sec], 1 [files processed], 0 [files errored], 1378.11[GFLOPS], filter:2.10723[sec], 2773.83[GFLOPS])

K:\w2x\waifu2x-converter-cpp\out\Release>
```
## After:

Silent:
```
K:\w2x\waifu2x-converter-cpp\out\Release>waifu2x-converter-cpp.exe -i 7KMAtWE.jpg -s
CUDA: GeForce RTX 2060
Processing file [1/1] "7KMAtWE.jpg": Done, took: 2.86705s total, file: 2.86705s avg: 2.86705s
process successfully done! (all:2.9237[sec], 1 [files processed], 0 [files errored], 1999.21[GFLOPS], filter:1.80589[sec], 3236.68[GFLOPS])
```
Verbose:
```
K:\w2x\waifu2x-converter-cpp\out\Release>waifu2x-converter-cpp.exe -i 7KMAtWE.jpg
CUDA: GeForce RTX 2060
Processing file [1/1] "7KMAtWE.jpg":
start process block (1/3,1/4) ...
.............
.............
.............
start process block (5/5,7/7) ...
Done, took: 4.21842s total, file: 4.21842s avg: 4.21842s
process successfully done! (all:4.27478[sec], 1 [files processed], 0 [files errored], 1367.34[GFLOPS], filter:2.12139[sec], 2755.32[GFLOPS])
```