@YukihoAA 2019-05-06
can we add skipped files count to file skipped?

check format function will skips some files but that count is not included final summary..