@DeadSix27 2019-05-06
New silent final log summary, with files skipped. (Examples at bottom).

Should we count the skipped files in the total too?

Currently it uses the final file list's count, which doesn't include skipped files: `int files_count = static_cast<int>(files_list.size());`

Making it say [4/6], with 3 files skipped, instead of  [7/9].

Both is correct, just matter or preference.

```
K:\w2x\waifu2x-converter-cpp\out\Release>waifu2x-converter-cpp.exe -i test -s -o test_out
CUDA: GeForce RTX 2060
We're going to be operating in a directory. dir:K:\w2x\waifu2x-converter-cpp\out\Release\test
Skipping file '.htaccess' for having an unsupported file extension (.htaccess)
Skipping file '9dd4d0fc35542950a82795a3ef306ceb.gif' for having an unsupported file extension (.gif)
Skipping file 'lolicondeyokatta2 - Copy (2).xyz' for having an unsupported file extension (.xyz)
Processing file [1/6] "312fbfbdb2af2c00bb7dae1cec1aeead.png": Done, took: 6s total, file: 1.15617s avg: 1.2128s
Processing file [2/6] "7KMAtWE.jpg": Done, took: 7s total, file: 2.76302s avg: 1.98879s
Processing file [3/6] "lolicondeyokatta2 - Copy.bmp": Done, took: 4s total, file: 0.45244s avg: 1.47723s
Processing file [4/6] "lolicondeyokatta2 - Copy.tif": Done, took: 2s total, file: 0.453708s avg: 1.22184s
Processing file [5/6] "lolicondeyokatta2.jpg": Done, took: 1s total, file: 0.465898s avg: 1.07104s
Processing file [6/6] "????????.jpg": Done, took: 0s total, file: 0.45539s avg: 0.968784s
Finished processing 6 files, took: 5.814542secs total, filter: 3.585004s. 3 files skipped, 0 files errored. [GFLOPS: 1926.793031, GFLOPS-Filter: 3125.078262]
```

Verbose:
```
.....
.....
start process block (3/3,3/3) ...
.....
Iteration #7(128->3)...(0.5766[ms], 3.88395[GFLOPS], 0.294443[GB/s])
total : 0.0223238[sec], 8.41839[GFLOPS]
Done, took: 0s total, file: 0.847104s avg: 1.58686s
Finished processing 6 files
Took: 9.523894secs total, filter: 4.574831s. 3 files skipped, 0 files errored. [GFLOPS: 1176.348598, GFLOPS-Filter: 2448.925218]

```