@DeadSix27 2019-05-06
New iteration formatting, see above comment for the other changes.

Now aligned and padded for cleaner reading/parsing.
```
total : 0.014[sec], 0393.78[GFLOPS]
start process block (3/3,3/3) ...
Iteration #1(  3-> 32)...(0.001[ms],    0.87[GFLOPS],    0.070[GB/s])
Iteration #2( 32-> 32)...(0.001[ms],    9.06[GFLOPS],    0.126[GB/s])
Iteration #3( 32-> 64)...(0.001[ms],   20.91[GFLOPS],    0.218[GB/s])
Iteration #4( 64-> 64)...(0.001[ms],   37.05[GFLOPS],    0.257[GB/s])
Iteration #5( 64->128)...(0.001[ms],   65.54[GFLOPS],    0.341[GB/s])
Iteration #6(128->128)...(0.001[ms],   80.14[GFLOPS],    0.278[GB/s])
Iteration #7(128->  3)...(0.000[ms],    4.56[GFLOPS],    0.346[GB/s])
total : 0.007[sec], 0026.12[GFLOPS]
Done, took: 0s total, file: 0.522s avg: 1.090s
Finished processing 6 files
Took: 6.539secs total, filter: 3.836secs; 3 files skipped, 0 files errored. [GFLOPS: 1713.30, GFLOPS-Filter: 2920.46]
```